Fix README to reflect actual behavior.
Originally I was going to completely ban globs but decided that
it would make more sense just to make the default be globs disabled.
That way the API behaves like how I want rimraf’s API to behave.
Globbing is a nice feature and convenient to have baked into rimraf
but something that should be explicitly opted into because of the
security concerns.