Pin rimraf to minor version for safety.
According to #1, rimraf can change behavior in non-backward-compatible
ways between minor versions and thus violates semver. Thus, we cannot
use a semver-friendly dependency on it.

As a result, will have to publish more versions of this package so
that people. For the most part, people should feel safe with a semver
dependency on rimraf-noglob as long as rimraf-noglob ensures that its
mechanism for disabling glob by default works with all minor versions
of rimraf it is published for.

Fixes #1.